requirements.txt
streamlit
openai
psycopg2-binary
sqlalchemy
python-dotenv
bcrypt
requests
